tpm: add single step functions to perform attestation
Split generation of the token based on the AK and the EK and the
client authentication.
In particular, make the new funcions agnostic on the channel making
use of a io.ReadWriter interface.
This way, the library could be used for the TPM attestation related
tasks, with no assumptions on the channel or the (extra) data not related
to the TPM attestation exchanged between the client and the server.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>